requirements.txt
openai
pyyaml
pandas
nltk
rouge-score
rich
